consumer/requirements.txt
google-cloud-pubsub
mysql-connector-python
pydantic
